+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added utility to get device name from hl-smi ([#232](https://github.com/Lightning-AI/lightning-Habana/pull/232))
+- Integrated Intel Neural Compressor for FP8 inference ([#235](https://github.com/Lightning-AI/lightning-Habana/pull/235))
+-
 
 ### Changed
 
@@ -28,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+- Deprecated support for Habana Quantization Toolkit. ([#235](https://github.com/Lightning-AI/lightning-Habana/pull/235))
 -
 
 ## [1.6.0] - 2024-06-28